Move generate func
